requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pyyaml==6.0.1
crewai>=0.100.0
python-dotenv==1.0.0